php arr - fixes
