+++ 18-intersect/Makefile
@@ -1,0 +1,6 @@
+console:
+	@php-console
+
+test:
+	@vendor/bin/phpunit --colors=always tests
+
