+++ 18-intersect/composer.json
@@ -1,0 +1,10 @@
+{
+  "autoload": {
+    "files": [
+      "src/Strings.php"
+    ]
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^11.4"
+  }
+}
