+++ 18-intersect/readme.md
@@ -1,0 +1,9 @@
+# src\Strings.php
+Реализуйте функцию checkIfBalanced, которая проверяет балансировку круглых скобок в арифметических выражениях:
+
+```php
+<?php
+
+checkIfBalanced('(5 + 6) * (7 + 8)/(4 + 3)'); // true
+checkIfBalanced('(4 + 3))'); // false
+```
